+++ example/lib/text/at_text.dart
@@ -4,13 +4,10 @@ import 'package:flutter/material.dart';
 
 class AtText extends SpecialText {
   AtText(TextStyle textStyle, SpecialTextGestureTapCallback onTap,
-      {this.showAtBackground = false, this.start})
+      {@required this.start})
       : super(flag, ' ', textStyle, onTap: onTap);
   static const String flag = '@';
   final int start;
-
-  /// whether show background for @somebody
-  final bool showAtBackground;
 
   @override
   InlineSpan finishText() {
@@ -19,33 +16,18 @@ class AtText extends SpecialText {
 
     final String atText = toString();
 
-    return showAtBackground
-        ? BackgroundTextSpan(
-            background: Paint()..color = Colors.blue.withOpacity(0.15),
-            text: atText,
-            actualText: atText,
-            start: start,
-
-            ///caret can move into special text
-            deleteAll: true,
-            style: textStyle,
-            recognizer: (TapGestureRecognizer()
-              ..onTap = () {
-                if (onTap != null) {
-                  onTap(atText);
-                }
-              }))
-        : SpecialTextSpan(
-            text: atText,
-            actualText: atText,
-            start: start,
-            style: textStyle,
-            recognizer: (TapGestureRecognizer()
-              ..onTap = () {
-                if (onTap != null) {
-                  onTap(atText);
-                }
-              }));
+    return SpecialTextSpan(
+      text: atText,
+      actualText: atText,
+      start: start,
+      style: textStyle,
+      recognizer: (TapGestureRecognizer()
+        ..onTap = () {
+          if (onTap != null) {
+            onTap(atText);
+          }
+        }),
+    );
   }
 }