+++ example/lib/text/my_special_text_span_builder.dart
@@ -1,5 +1,4 @@
 import 'package:extended_text_library/extended_text_library.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'at_text.dart';
@@ -7,19 +6,7 @@ import 'dollar_text.dart';
 import 'emoji_text.dart';
 
 class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
-  MySpecialTextSpanBuilder({this.showAtBackground = false});
-
-  /// whether show background for @somebody
-  final bool showAtBackground;
-  @override
-  TextSpan build(String data,
-      {TextStyle textStyle, SpecialTextGestureTapCallback onTap}) {
-    if (kIsWeb) {
-      return TextSpan(text: data, style: textStyle);
-    }
-
-    return super.build(data, textStyle: textStyle, onTap: onTap);
-  }
+  MySpecialTextSpanBuilder();
 
   @override
   SpecialText createSpecialText(String flag,
@@ -28,13 +15,12 @@ class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
       return null;
     }
 
-    ///index is end index of start flag, so text start index should be index-(flag.length-1)
+    // index is end index of start flag, so text start index should be index-(flag.length-1)
     if (isStart(flag, AtText.flag)) {
       return AtText(
         textStyle,
         onTap,
         start: index - (AtText.flag.length - 1),
-        showAtBackground: showAtBackground,
       );
     } else if (isStart(flag, EmojiText.flag)) {
       return EmojiText(textStyle, start: index - (EmojiText.flag.length - 1));